+++ CHANGES.rst
@@ -6,6 +6,7 @@ Version 0.3.4
 =============
 
 - weighted SWI does not calculate QFLAG by default (which should be done only daily). Needs proper implementation in the future.
+- weighted SWI rectifies overshooting due to weighting (--> 0% > SWI < 100%)
 
 Version 0.3.3
 =============